Remove redundant mapArray utility function
We previously used the vararg overload of Markers.aggregate to pass
multiple markers when using the LocationAwareLogger, using the spread
operator with an array. However, the spread operator creates an
additional copy of the array. We can avoid this by using a different
overload of Markers.aggregate, that takes a Collection. We then no
longer need the mapArray utility to map an array, since we now instead
use a List.