Replace use of "marker" with "log field"
"Marker" is an SLF4J-specific term, that was not originally designed for
adding key-value data to logs in the way that we use it. Markers are
more like "tags" on a log. But the logstash-logback-encoder library,
which we use to attach structured key-value data to logs, use markers
to pass this data.

Initially, I thought that using the term "marker" made sense, since that
is the underlying concept we use from logstash-logback-encoder. But this
confuses the logstash-specific key-value marker concept with the more
general SLF4j concept of a marker. In addition, no other languages that
I know of use the term "marker" to describe key-value data added to
logs.

So with this in mind, I've decided to instead use the term "log field".
After all, the way we use markers internally in the library is just to
add key-value pairs to the log output - i.e., fields.