Move call to LogEvent.log out of LogBuilder
The previous implementation made it so that both LogBuilder and LogEvent
were involved in logging. This turned out to not work well with caller
information in certain SLF4J implementations, namely slf4j-jdk14, which
can apparently only exclude 1 class from the stack when calculating
caller boundary.

By moving the call to LogEvent.log out to Logger.log, which is inline,
only the LogEvent implementation class is involved in the log, making
it work with slf4j-jdk14.