Revert "Use SLF4J LoggingEventBuilder to support Log4j key/values"
This reverts commit f95783fe843d400bc4e7af1754e355dffc9b3cf3.

log4j-slf4j2-impl only supports string -> string key/value pairs, which
means that all our log fields would be escaped. This, in addition to
the more complex solution we had to implement for Slf4jLogEvent to use
the LoggingEventBuilder, has led me to conclude that this is not worth
it at the moment. Once log4j-slf4j2-impl improves their key/value
handling, we can revisit this.