Revert "Strip platform file extension from logger names"
The idea behind the last commit was to omit the platform extension
recommended for Kotlin Multiplatform file names (e.g. "jvm" in
`Example.jvm.kt`) from loggers created in such files. However, this
could lead to false positives for classes/files with underscores in
their name. An example of when you would have this is when using the
Flyway migration library, where so-called "Java-based migrations" are
classes with underscores in the name.

So the feature is not really feasible as intended. The downside is
really just that we don't strip away the trailing "_jvm" suffix from
synthetic class names, which is not the end of the world. Users can
customize the logger name by using the `getLogger` overload that takes
a custom name, or use `@file:JvmName` to change the name of the
generated class for a file.

This reverts commit 054f24a1e349a70bc0c8992acead032c77c9935a.