Implement custom Logback ThrowableProxy to hide LoggingContextProvider
We use a LoggingContextProvider exception to smuggle logging context
through suppressed exceptions. But we don't want this exception to
actually show up in the stack trace in the logs, since it's an
implementation detail of the library. So we implement the
IThrowableProxy interface from Logback, which lets us exclude this
exception from the suppressed exceptions.