Remove crossinline modifier from Logger lambda arguments
This keyword complicated the signature of Logger methods, without
providing much real benefit. Users making non-local returns inside a
log lambda isn't a _problem_ for the library, and was rather meant to
prevent users from doing something unintended. But we deem it a higher
cost that users must learn of the non-trivial `crossinline` concept to
understand the methods on `Logger`.