Use literal instead of empty String constructor for JSON_FIELD_SENTINEL
The empty String constructor is not supported on all Kotlin platforms
(see https://kotlinlang.org/api/core/kotlin-stdlib/kotlin/-string/-string.html).
Since we use the presence of JSON_FIELD_SENTINEL or a `null` to mark if
a field is JSON, we don't really need this to be a unique instance, just
a non-null String - so we can just use a string literal instead.