Remove overrides of printStackTrace on LoggingContextProvider
We previously had overrides of these methods to be no-ops. But this
could lead to confusing results if these methods were used to append to
log output, since you would then see "Suppressed: " followed by nothing.

The default implementation of `printStackTrace` includes the `toString`
representation of the exception (class name + message), which is OK. No
stack trace will be printed, since we override `fillInStackTrace` - and
that's what we want.