Remove uses of @JvmStatic
Previously, we used @JvmStatic as a potential performance optimization,
thinking that it would avoid the indirection of going through the
companion object. This was a false assumption: @JvmStatic annotations
instead generate an _additional_ function, which then calls the
corresponding function on the companion object. This just generates
another layer of indirection, and bloats the bytecode.

So now we instead use top-level functions, which compile to actual
static methods.